--- src/main/java/gift/exception/oauth/OAuthException.java
@@ -1,12 +1,0 @@
-package gift.exception.oauth;
-
-public class OAuthException extends RuntimeException {
-
-    public OAuthException(String message) {
-        super(message);
-    }
-
-    public OAuthException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
